完成

+++ online-course-api/src/main/java/com/course/api/entity/CourseVideo.java
@@ -1,14 +1,15 @@
 package com.course.api.entity;
 
-import com.baomidou.mybatisplus.annotation.TableName;
 import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.TableId;
-import java.time.LocalDateTime;
 import com.baomidou.mybatisplus.annotation.TableField;
-import java.io.Serializable;
+import com.baomidou.mybatisplus.annotation.TableId;
+import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
+
+import java.io.Serializable;
+import java.time.LocalDateTime;
 
 /**
  * <p>
@@ -72,7 +73,7 @@ public class CourseVideo implements Serializable {
      * 视频长度
      */
     @TableField("video_length")
-    private Integer videoLength;
+    private Long videoLength;
 
     /**
      * 排序
